% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weak_dataset.R
\name{generate_weak_dataset}
\alias{generate_weak_dataset}
\title{Generate Weak Dataset}
\usage{
generate_weak_dataset(
  n = 100,
  prop_cluster = c(0.5, 0.5),
  cluster_functions = list(fun_one = function(x) pweibull(x, shape = 6, scale = 0.5,
    lower.tail = FALSE), fun_two = function(x) 0),
  num_distances = list(Rural = 15, Urban = 15),
  rdist_generator = list(Rural = function(x) rbeta(x, 1, 1), Urban = function(x)
    rbeta(x, 1, 1)),
  prop_exposed = 0.95,
  Z = cbind(1, rbinom(n = n, 1, 0.5)),
  delta = c(26, -0.5),
  seed = NULL
)
}
\arguments{
\item{n}{number of observations}

\item{prop_cluster}{vector of cluster proportions}

\item{cluster_functions}{list of length \code{num_clusters}}

\item{prop_exposed}{proportion of observations with exposure to spatial features}

\item{Z}{matrix of standard regression coefficients - assumes first column is an intercept}

\item{delta}{vector of regression coefficients}

\item{seed}{random number generator seed}

\item{mean_num_distances}{mean number of simulated distances to include}
}
\description{
Generates clustered nonlinear spatial exposure effects in a linear model where distance distribution and count is correlated with the cluster function
}
