% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fit_multiple.R
\name{sim_fit_multiple}
\alias{sim_fit_multiple}
\title{Simulation Study}
\usage{
sim_fit_multiple(
  num_sim = 10L,
  K = 50,
  fix_alpha = FALSE,
  iter_max = 5000,
  burn_in = 3000,
  chains = 4,
  dataset_function
)
}
\arguments{
\item{num_sim}{number of simulations to run}

\item{....}{other arguments for \code{\link{generate_dataset}}}
}
\description{
Simulation Study
}
